+++ app/src/main/java/com/numero/material_gallery/components/button/MaterialButtonToggleGroupActivity.kt
@@ -1,4 +1,4 @@
-package com.numero.material_gallery.activity
+package com.numero.material_gallery.components.button
 
 import android.content.Context
 import android.content.Intent
@@ -9,10 +9,10 @@ import androidx.appcompat.app.AppCompatActivity
 import com.numero.material_gallery.R
 import com.numero.material_gallery.fragment.ThemeInfoBottomSheetDialog
 import com.numero.material_gallery.repository.ConfigRepository
-import kotlinx.android.synthetic.main.activity_radio_button.*
+import kotlinx.android.synthetic.main.activity_material_button_toggle_group.*
 import org.koin.android.ext.android.inject
 
-class RadioButtonActivity : AppCompatActivity(R.layout.activity_radio_button) {
+class MaterialButtonToggleGroupActivity : AppCompatActivity(R.layout.activity_material_button_toggle_group) {
 
     private val configRepository by inject<ConfigRepository>()
 
@@ -46,6 +46,6 @@ class RadioButtonActivity : AppCompatActivity(R.layout.activity_radio_button) {
     }
 
     companion object {
-        fun createIntent(context: Context): Intent = Intent(context, RadioButtonActivity::class.java)
+        fun createIntent(context: Context): Intent = Intent(context, MaterialButtonToggleGroupActivity::class.java)
     }
 }
