+++ app/src/main/res/layout/activity_checkbox.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.CheckboxActivity">
+    tools:context=".components.selection.CheckboxActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
